Fix perfomance issue with Dual Strike
The 'BondmateDualStrikeFollowup' ability was suffering from the same
lack of gameplay visibility constraints as Rapid Fire and Chain Shot,
leading to a lot of aim checks at the start of the mission and when
moving a soldier.

This adds a gameplay visibility condition to that ability. Note that it
doesn't use the `MULTI_SHOT_ABILITIES` config because the alternate
mechanic for fixing the performance issue won't work. Dual Strike
requires 'BondmateDualStrikeFollowup' to be available before
'BondmateDualStrike' is activated.