Fix to some Hack Reward Abilities descriptions
Vanilla uses plain text for aim/crit/dodge values in Hack Reward effects descriptions, even though appropriate tags already exist and are used in the hack screen descriptions. Since LWOTC changes default values, this discrepancy becomes visible. This replaces plain text with the tags. LWOTC Hypnography value is the same as vanilla, so the description was technically correct, but better fix it as well, in case it might change in the future.