Make Alien Facilities controller compatible (#1379)
Modify UIStrategyMapItem_AlienNetworkComponent_LW so that Alien
Facilities are clickable with a controller.

Description : A while back, UIStrategyMapItem_Mission_LW was modified
so that strategy map items could be 'clicked' and would interact
appropriately with a controller. Unfortunately, I didn't realize that
the alien facility had its own special strategy map item class,
UIStrategyMapItem_AlienNetworkComponent_LW.

Consequently, I have transferred all of the changes I made to
UIStrategyMapItem_Mission_LW into
UIStrategyMapItem_AlienNetworkComponent_LW.

Here is a general discussion of the changes made from my commits for
UIStrategyMapItem_Mission_LW

=========================

Modifies UIStrategyMapItem_AlienNetworkComponent_LW --> InitMapItem()

When using a controller, an infiltrating mission displays a help icon,
StrategyScanButton --> consoleHint, to the left of the haven information
button. The problem is that it overlaps the icon representing the
infiltrating squad. Consequently, the best way to deal with it, is to
null it.

=========================

Modifies : UIStrategyMapItem_AlienNetworkComponent_LW -->
 UpdateFromGeoscapeEntity()

If an infiltrating mission map item is selected, make sure the
infiltration percentage string is black. If an infiltrating mission map
item is not selected, make sure the infiltration percentage string is
its normal blue. If this is not done, the text will 'disappear' into
the background colour.

Additionally, within this function, update the map item tooltip
information when appropriate.

=========================

Modifies UIStrategyMapItem_AlienNetworkComponent_LW --> OnMouseEvent()
 and OnUnrealCommand().

Originally, when a controller user clicked on the A button while a map
item mission was highlighted, it would open a normal mission screen
even if the mission was currently undergoing infiltration. As a result,
OnUnrealCommand() had to be extended so as to call the same code as
found in OnMouseEvent() --> FXS_L_MOUSE_UP. This code calls
OpenInfiltrationMissionScreen(), which pushes an infiltration screen
onto the screen stack if infiltration is occurring.

=========================

Added : UIStrategyMapItem_AlienNetworkComponent_LW --> GetToolTipText,
 UpdateTooltip, and GenerateTooltip

Provides the opportunity for controller users to display mission
tooltips next to the mission map item. Formerly, they would only be
accessible to mouse and keyboard users via the mission icons on the
bottom icon bar.