Fixing Repression activity
As reminder of what repression is, this is the description from XComLW_Activities.ini: "Repression works only in uncontacted regions; it is simulating the Resistance without XCOM's assistance. It rolls several d100s once the cooldown is over."

As far as I can tell, and as far as I can remember this activity never worked, even in the original LW2. The reason for this is that RepressionComplete function seeks for RegionState only in the NewGameState, and immediately fails if it doesn't find it. It is unlike all other activities which seek XCOMHistory in such case. I am fuzzy on the details for  why activities try to find RegionState object in NewGameState at all, since apparently it always fails, but simply doing what other activities do should work for Repression as well.
The second problem was that the iRoll variable simply wasn't assigned value  properly, thus it would've had value of 0 for all checks.
Third  problem was the condition check for REPRESSION_REBEL_LOST_CHANCE was inverted for reasons unknown. Probably a typo.

With this fixes this activity seem to work as seemingly intended. However, I should reiterate that I think it never ever worked, and thus might had never been playtested, balance-wise. Therefore, I am not completely certain that this fix should even be made, though I am leaning to it, since I like the idea behind the mechanic.
Looking at the chances of events happening, here is the average change of  one repression activity: -0.02 Alert, -0.03 Vigilance*, =0 Rebels**, which should be multiplied by number of uncontacted regions to get average world-wide change per duration period.
(* slightly more, since vigilance decreases of REBEL_LOST events only happen if removed rebel was non-faceless, but it is hard to adjust for)
(** the percentage of faceless will rise over time, since events can add both faceless and normal rebels, but are removing only normal ones)

Since the Vigilance decrease of REPRESSION_REBEL_LOST and REPRESSION_2ND_REBEL_LOST events were not described as results in the ini file, I am not sure if Grobobobo had considered it when he made the last changes to the percentage chances on Nov 21, 2020. One might want to adjust the chances.