Fix Repression activity
As reminder of what repression is, this is the description from
XComLW_Activities.ini:

  "Repression works only in uncontacted regions; it is simulating the
  Resistance without XCOM's assistance. It rolls several d100s once the
  cooldown is over."

As far as I can tell, and as far as I can remember this activity never
worked, even in the original LW2. The reason for this is that
RepressionComplete function seeks for RegionState only in the
NewGameState, and immediately fails if it doesn't find it. It is unlike
all other activities which seek XCOMHistory in such case. I am fuzzy on
the details for  why activities try to find RegionState object in
NewGameState at all, since apparently it always fails, but simply doing
what other activities do should work for Repression as well.

The second problem was that the iRoll variable simply wasn't assigned
value properly, thus it would've had value of 0 for all checks.

Third problem was the condition check for REPRESSION_REBEL_LOST_CHANCE
was inverted for reasons unknown. Probably a typo.

Lastly, rebels were not being removed if the random "pick" of a rebel
happened to be a Faceless. This reduced the overall chance of Repression
actually removing a rebel, which isn't intuitive from the config. You
would expect REPRESSION_REBEL_LOST_CHANCE to be the actual chance if
there is at least one non-Faceless rebel in the haven.

Resolves #1411.