Remove Indomitale and Terrorize from templar's Ghosts
No one bothered to add these two abilities to the exclusion list yet, so I might as well.
Indomitable should be removed because ghosts should not be able to raise  their focus.
Terrorize should be removed because ghosts don't have Volt in the first place and so it is cleaner this way.