Add conditional check on $cookedMaps to fix build error
When pulling the repo down for the first time and building, I
received a build error relating to these sorting methods on the
cookedMaps variable in this script.

The list had a Length of 0 and when sorted, the variable's type was changed.
Further along in the build I would then see the error:
"The property 'Length' cannot be found on this object."

This fixed the issue by preventing the sort of an empty list and the
unintended type-change which occurs.