Disable CA soldier capture risk
The flat risk was unengaging and too much of a disincentive to running
covert actions, since you generally want to run them with high(ish)
ranked soldiers. The Intel cost to remove the risk allowed some control,
but it ended up being a situation of paying the cost or not running the
covert action because captures hurt so much. And Intel is often tight as
it is.