Add Overwatch to the DoubleTapAbilities whitelist
Classes other than Sharpshooter can get Double Tap ability with XCom row, but they won't be able to use it for overwatching despite the description explicitly saying they should be able to. Normal 'Overwatch' ability isn't on the list of allowed abilities which you can spend bonus action point on. This fixes it.