Replace XComLWTuple with LWTuple in OnShouldShowPsi
As far as I can understand OnShouldShowPsi function exists to provide support for Detailed Soldiers List mod (which I use), where the 'DSLShouldShowPsi' event is triggered in the first place. However, right now this function spam the logs with 'Warning: OnShouldShowPsi event triggered with invalid event data.' messages. It seems to me that it is because the DSL sends LWTuple-s instead of XComLWTuple objects. Changing it does fix it.

Visibly, this change replaced the display of ComInt stat for rookies with psi offense values. It may or may not be desirable.

I don't know why OnShouldShowPsi expected XComLWTuple in the first place. Was it changed in DSL at some point? Given that there is some history with integrating/de-integrating DSL, I am also unclear on whether this function is even still meant to be there. In any way, even if it is just a remnant, then it should be removed altogether to not spam the logs.