Fix Hyper Reactive Pupils PCS
As described in #1447 HRP bonus currently applies after any unit misses, instead of just unit with HRP.
Now, technically speaking, one could read the description such that "a miss"  in the HRP description refers to any miss whatsoever. And the fact that no check that the "miss was made by the unit with the PCS" was even attempted in the code could mean that it was intended this way. However, I think it would be very bizarre thematically, and I believe no player ever thought that it would be how it works after reading its description.

Thus, the check for the miss to be made by the unit with HRP pcs is added.

Additionally, I added several entries to the HRP-related SHOTFIRED_ABILITYNAMES list, which do probably belong there. There are probably other fitting abilities missing form this list, but I couldn't think of any immediately.