Make Locked On apply to primary weapon only
Currently Locked On bonuses can only be triggered by primary weapon shots, but applies to non-primary weapons (like pistols and sawn-off shotguns) as well.
This fixes this inconsistency by making it both triggered by and applied to primary weapon shots only.  It may be argued that it would be better to have it other way around and work with both. It also may be argued that it should not be fixed at all since, I believe, it is how it worked in LW2 as well. But I think this way is preferable overall, especially considering LockedOn values were buffed compared to LW2.