Fix Lightning Reflexes
There are two problems with how LR counter (which is responsible for  effect reduction) is currently calculated:
First is that the listener does not check that the defending unit with LR and the unit with the LR whose listener was called is the same unit. In practice it means that if there multiple units with LR present on the map (let's say there are 5 Viper Elites somewhere far away), then the counter would be increased by 1 for each such unit after the first attack is evaded. Next attempt to run another overwatch this turn would thus have 0 reduction effect from LR. Which is not how it supposed to work. This change adds this check.

Second problem is that the GetToHitAsTargetModifiers is checking for attack to be bReactionFire, but the IncomingReactionFireCheck which increases the LR counter check that the attack is on the LR_REACTION_FIRE_ABILITYNAMES whitelist. These checks are not the same.  For example, lost' grapple attack are not on this whitelist even though it is reaction fire. It means in effect that unit with LR can dodge any number of grapple attacks in a single turn with -100 to-hit modifier, as the counter would not increase. I believe there are other examples of attacks which are reaction fire but are not on that list. This change removes this whitelist and instead checks for attack to be reaction fire instead, same as to-hit mod function. I am not fully sure why LR would use this list in the first place or if there are any problems which could be caused by deleting it from configs, but as I see this list have no real purpose and is prone to be outdated.