Chosen-related text fixes
Fixes capitalization and parameterizes some abilities' values. Fixes missing text  for  "Summon Followers" and wrong +crit bonus for "Wrath Against Unworthy" (it is actually 100).

Not fixed: XCom version of Darkclaw pistol have Fatality ability associated with it  When viewed in inventory it will erroneously state that it will apply to primary weapon. It is because 'WeaponName' tag handler will fail to find what it is associated with and will resort to generic primary weapon string. I couldn't find an easy way to fix it.