Fix Combat Rush PCS (#1455)
Currently Combat Rush effect from PCS can only be triggered by kills where there is "primary target". So, for example, if you kill an enemy using grenades, then it won't trigger. Most likely the cause of this: #1401.

For some reason current implementation tries to catch kills in PostAbilityCostPaid function. It seems weird when there are "UnitDied" and "KillMail" events already. Though, I am not sure what are the exact practical differences between the two.
This replaces PostAbilityCostPaid approach with KillMail listener. Note, that
BroadcastCombatRush still left with the trigger for 'CombatRush'-named  event, even though nothing call it now. I was not sure whether to remove it as obsolete, or to keep in case something else will try to use 'CombatRush' event.