Fix to InfiltrationHackBonus change (#1450)
The config change of setting InfiltrationHackBonus value to 0 was put in the wrong ini file and does not take the effect.
A trivial fix, which you even might have made locally already. It is here mainly in case its issue was accidentally skipped over and so to not be forgotten about.