Make Locked On not lose the bonus on non-offensive abilities (#1449)
* Make Locked On not lose the bonus on non-offensive abilities.

Currently ranger's Locked On effect is lost when using any non-offensive ability or using non-primary weapon. This behavior differs from original LW2 and, I think, feels very janky.
From what I can tell, here is how it happened. Originally Locked On relied on XComGameState_Effect_LastShotDetails class to keep track of who was shot last. However, this class' event listener does not make the check that the unit using an ability is the same as the unit which have relevant effect (which is unclear as to why for me, but I suspect it is either the difference in how it is triggered in vanilla or how it was subscribed to in LW2).  Which lead to there being an issue where locked on was cleared by other soldiers using their abilities. That issue was fixed there https://github.com/long-war-2/lwotc/pull/621 by replacing Locked On listener to vanilla ZeroIn listener.  However, said ZeroIn listener differs from the original RecordShot in certain details. Namely, it resets its counters if the unit uses any non-offensive ability or uses non-primary weapon.

The fix is to use custom listener, which mostly copied from ZeroIn, but with resetting counters removed.

Another difference from RecordShot is that it used SHOTFIRED_ABILITYNAMES whitelist of abilities which could apply Locked On. Now it uses ZeroIn criteria which is any targeted attack with primary weapon. I couldn't think of an ability which would be in one set but not another, but there could be some.

* Make Locked On apply to primary weapon only

Currently Locked On bonuses can only be triggered by primary weapon shots, but applies to non-primary weapons (like pistols and sawn-off shotguns) as well.
This fixes this inconsistency by making it both triggered by and applied to primary weapon shots only.  It may be argued that it would be better to have it other way around and work with both. It also may be argued that it should not be fixed at all since, I believe, it is how it worked in LW2 as well. But I think this way is preferable overall, especially considering LockedOn values were buffed compared to LW2.

* Make LockedOn values not cleanup every turn

I forgot that the default for SetUnitFloatValue is eCleanup_BeginTurn. Which makes locked on target go away every turn. This should fix it.