Disable Show of Force sit rep
It doesn't work very will in LWOTC because the increase in pod size
doesn't apply and the selection of units for the ADVENT pods appears to
be based on vanilla spawns. This means you can get all-M2 troopers in a
pod at FL7, which is a huge difficulty spike. And yet it's pretty
trivial at later force levels until the FL for M2 troopers is hit.

Lastly, it confusingly shares a name with the LW2 Show of Force dark
event that increases the alert level of missions.