Clarify Blood Trail anti-Dodge mechanic in l10n
The description of the ability talked about reducing the target's Dodge,
but this implied that the effect was persistent and would apply for
other soldiers shooting the target. It could also imply that the Dodge
could go negative.

The description now says that Blood Trail simply ignores up to 40 Dodge,
which is hopefully less ambiguous.