Show granted items on completed research (#3)
This change adds a listener for the 'ResearchReport' event that forces a
popup for any items that are granted by the corresponding research that
just completed.

I have also added a listener that fixes the weapon images used for the
"item unlocked" popups after research. However, this problem is quite
pervasive (see #285) so it's only a very partial fix.

Fixes #3.