Add min crate recovery objective to Supply Extraction
I've repurposed the old "kill all enemies" display objective so that it
displays the minimum number of crates recovered for a successful
mission. The minimum crate count is only displayed for English right
now, but I have updated the other languages to show the equivalent of
"Mark and extract crates".

This should make it clearer to players that they need to hit a minimum
count to successfully complete the mission.