Add framework for mission detection modifiers
LW2 had a 'GetActivityDetectionIncomeModifier' event that allowed mods
to modify the detection income for any activity. I have leveraged this
to create a framework for mission detection modifier templates. These
provide different ways of influencing mission detection in a given
outpost.

The main advantage of these detection modifiers is that they can
activate based on the activity that is being detected. So you can grant
bonuses for one set of activities but not another.

This commit adds three permanent detection modifiers to outposts that
give boosts to the detection of particular activities based on which
jobs rebels are on. This allows rebels on recruit and supply to have an
influence over mission detection, not just those on intel.

I have also added a template for boosting liberation detection, but it's
not used yet. It needs to be added as a reward that could then be
granted from covert actions or objective hacks, or things like that.