Reveal Avatar progress on first pip of doom
This ensures players have a reasonable idea of how well they're doing,
as a good campaign will have slower pip progression than one that does
nothing. This also fixes the problem with the Avatar project not
completing when it hasn't been revealed (so players end up with more
than 20 pips of doom when the finally reveal it).