Fix mission detection modifiers
The template creation class was not extending the correct class which
meant the templates weren't being created. In addition, the game state
detection modifier objects were being stored in the outpost, which left
that array out of date. It now stores `StateObjectReference`s.

Lastly, the percentage modifiers were applying to the accumulated
mission income rather than the base income gained during a single
update. This meant that the bonuses got larger and larger quite
quickly!