Mission Covert Op 2: mini troop column
in the vein of ambushing a scouting party, this covert op locates a small full salvage mission to get some corpses.  Mission is capped at STR 4 no matter what region it spawns in, with a 5-6 day infil timer and cap of 6 soldiers to infiltrate.  Currently has standard "medium" settings for covert ops.