Permanent Advent Firewalls Dark Event from 100% to 33% of mechanicals
Request to change Advent Firewalls - Permanent Dark Event to affect only 33% of encountered robots, as seen on wiki.  

Secondary to the fact that it's not fun gameplay, it's too harsh.  Penalties for failed hacks is already very high.  

Perhaps a nerf to Failsafe would do instead, having it only prevent failed hack penalties a % of the time (75%, for example)