Improve hit chance display of guaranteed crit abilities
For abilities that force crits, now display the entire bar as crit and override the hit/graze calc display.  This is only a visual change and does not affect the behavior of the ability because of how the bHitsAreCrits flag works.  Related to a recent Highlander change that also adds crit chance to these abilities too.