The great Lib3 bugfix/tweak
Fixed the "Lib3 spawns with max timer after first detection"
Instead, Lib3 now has a scaling bonus that increases with the number of times you detect/let it expire, currently set to 15% increase in detection chance per time detected.

After first detection is 1.15x, after 2nd it's 1.32x, after 3rd it's 1.52x, and so on.  This system fixes the bug but in compensation, makes it progressively easier to detect each time.