Tweak reward soldier generation
after the previous bugfix, turns out that changed how reward soldiers were generated. Previously, it used the base game SelectNextSoldierClass which chose soldiers it thinks you "needed"  Current state is fully random.  This changes it to something in between, building the full class deck, then adding the needed classes to it as well, so they are weighted more.