Map shuffling during mission generation
the base game plot selection system only marks plots as "used" when you load into a mission, so when multiple missions spawn before a mission is played, the deck system will supply the same map multiple times as the "used the least" option, thus missions spawned at the same time are likely to use the same plot if said plot is valid for both missions.

This removes plots that already are in use from the list of valid plots before LW checks them for sitrep validity.