experimental pod job tweaks stuff + sniper defense tweaks
* Pod Manager now gets updated XCOM position by Sniper Defense alerts (if the unit survives).
* All pod job assignments minimum turns increased by 1 since the bugfix to pods freezing mean they would move 1 turn sooner.
* tweaked the turn timers for HQ assault pod job activations. This should spread stuff out more.
* Removed the legend-only extra flanking job from supply raids.
* Removed chosen supply raids from the HQ/big mission jobs lists. - should mean less enemies rush XCOM.
* Yellow alert pods using sniper defense code will now try to overwatch or hunker for second actions when visible, instead of just blindly moving towards XCOM.

Overall, this should improve the feeling of the slower long missions and reduce the amount of times XCOM gets rushed by the entire map on HQs and Supply Raids. However, this also does fix a loophole that previously allowed XCOM to exploit what happens when the pod manager never gets an actual position for xcom and instead vectors pods relative to the 0,0,0 location on the map.

I have tested the updated behaviors in both GOPs, supply raids, and HQs, but things still might be weird because it's XCOM AI